+++ apps/app/src/features/growi-plugin/client/components/index.ts
@@ -1,0 +1,1 @@
+export * from './GrowiPluginsActivator';
