+++ apps/app/src/features/growi-plugin/client/utils/growi-facade-utils.ts
