--- apps/app/src/features/growi-plugin/components/index.ts
@@ -1,1 +1,0 @@
-export * from './GrowiPluginsActivator.client';
